# static/openapi.yaml
openapi: 3.0.3
info:
  title: Api-SportMatch
  description: >
    API REST de SportMatch. Fase 2: autenticación, usuarios y administradores.
    Fase 3 (en construcción): deportes, cuestionario de registro, eventos,
    inscripciones y valoraciones, organizadores y validación, administración
    y soporte. Ver /docs/entidades_completas.md en el repositorio para el
    catálogo completo de las 89 entidades del modelo.
  version: 0.5.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/deportes:
    get:
      summary: Listar catálogo de deportes (público; ?incluir_inactivos=1 para admin)
      tags: [Deportes]
      security: []
      responses:
        "200": { description: Lista de deportes }
    post:
      summary: Crear un deporte (solo admin)
      tags: [Deportes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, example: "Running" }
                categoria: { type: string, example: "resistencia" }
      responses:
        "201": { description: Deporte creado }
        "409": { description: Ya existe un deporte con ese nombre }
  /api/deportes/{deporte_id}:
    get:
      summary: Obtener un deporte
      tags: [Deportes]
      security: []
      parameters:
        - { name: deporte_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Deporte encontrado }
    put:
      summary: Actualizar un deporte (solo admin)
      tags: [Deportes]
      parameters:
        - { name: deporte_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Deporte actualizado }
    delete:
      summary: Desactivar un deporte (borrado suave, solo admin)
      tags: [Deportes]
      parameters:
        - { name: deporte_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Deporte desactivado }
...
  /api/eventos:
    get:
      summary: Listar eventos (público; por defecto solo estado=publicado). Filtros - Query params - estado, tipo, deporte_id, ciudad_id, q, page, per_page
      tags: [Eventos]
      security: []
      responses:
        "200": { description: Lista paginada de eventos }
    post:
      summary: Crear un evento, opcionalmente con deportes/requisitos/sedes/fechas/categorías+boletos anidados (solo admin)
      tags: [Eventos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [titulo, tipo]
              properties:
                titulo: { type: string }
                tipo: { type: string, example: "carrera" }
                dificultad: { type: string, enum: [principiante, intermedio, avanzado, elite] }
                descripcion: { type: string }
                deportes: { type: array, items: { type: integer } }
                requisitos:
                  type: array
                  items: { type: object, properties: { descripcion: { type: string }, orden: { type: integer } } }
                sedes:
                  type: array
                  items:
                    type: object
                    properties:
                      nombre: { type: string }
                      ciudad_id: { type: integer }
                      latitud: { type: number }
                      longitud: { type: number }
                fechas:
                  type: array
                  items: { type: object, properties: { inicia_en: { type: string, format: date-time }, termina_en: { type: string, format: date-time } } }
                categorias:
                  type: array
                  items:
                    type: object
                    properties:
                      nombre: { type: string }
                      distancia_km: { type: number }
                      cupo_total: { type: integer }
                      boletos:
                        type: array
                        items: { type: object, properties: { tipo: { type: string }, precio: { type: number } } }
      responses:
        "201": { description: Evento creado (borrador) con todo lo anidado }
  /api/eventos/{evento_id}:
    get:
      summary: Detalle de un evento (con deportes/requisitos/sedes/fechas/categorías/boletos)
      tags: [Eventos]
      security: []
      parameters:
        - { name: evento_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Evento }
    put:
      summary: Actualizar un evento, incl. cambiar `estado` (borrador/publicado/cancelado/finalizado) (solo admin)
      tags: [Eventos]
      parameters:
        - { name: evento_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Evento actualizado }
    delete:
      summary: Eliminar un evento (borrado suave, solo admin)
      tags: [Eventos]
      parameters:
        - { name: evento_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Evento marcado como cancelado/eliminado }
  /api/eventos/slug/{slug}:
    get:
      summary: Detalle de un evento por su slug
      tags: [Eventos]
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Evento }
  /api/eventos/{evento_id}/categorias:
    post:
      summary: Agregar una categoría a un evento existente (solo admin)
      tags: [Eventos]
      parameters:
        - { name: evento_id, in: path, required: true, schema: { type: integer } }
      responses:
        "201": { description: Categoría creada }
  /api/eventos/categorias/{categoria_id}/boletos:
    post:
      summary: Agregar un tipo de boleto a una categoría (solo admin)
      tags: [Eventos]
      parameters:
        - { name: categoria_id, in: path, required: true, schema: { type: integer } }
      responses:
        "201": { description: Boleto creado }
...